chore: move exec struff into its own file
